% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ANALYSE_BASELINE.R
\name{ANALYSE_BASELINE}
\alias{ANALYSE_BASELINE}
\title{Create baseline analysis dataset.}
\usage{
ANALYSE_BASELINE(
  DISEASE_THEME = "",
  DATA_DM,
  DATA_IN = NULL,
  DATA_LB = NULL,
  DATA_MB = NULL,
  DATA_MP = NULL,
  DATA_RP = NULL,
  DATA_SA = NULL,
  DATA_TS = NULL,
  DATA_VS = NULL,
  DM_VARS = NULL,
  LB_VARS = NULL,
  MB_VARS = NULL,
  MB_VARS_SPEC = NULL,
  MP_VARS = NULL,
  IN_VARS = NULL,
  RP_VARS = NULL,
  SA_VARS = NULL,
  VS_VARS = NULL,
  IN_DUR = FALSE,
  IN_TIME = FALSE,
  SA_DUR = FALSE,
  SA_TIME = FALSE,
  MP_TESTCD = "LENGTH"
)
}
\arguments{
\item{DISEASE_THEME}{The name of the disease theme being analysed. Character
string. Default is empty (selects base variables). Select from: "MALARIA",
"VL" or "EBOLA". If selection is missing or misspelt, then the base
variables will be used.}

\item{DATA_DM}{The DM domain data frame, as named in the global environment.
Required.}

\item{DATA_IN}{The IN domain data frame, as named in the global environment.}

\item{DATA_LB}{The LB domain data frame, as named in the global environment.}

\item{DATA_MB}{The MB domain data frame, as named in the global environment.}

\item{DATA_MP}{The MP domain data frame, as named in the global environment.}

\item{DATA_RP}{The RP domain data frame, as named in the global environment.}

\item{DATA_SA}{The SA domain data frame, as named in the global environment.}

\item{DATA_TS}{The TS domain data frame, as named in the global environment.}

\item{DATA_VS}{The VS domain data frame, as named in the global environment.}

\item{DM_VARS}{Specify additional variables to be included in the output
dataset. Character string. Use column names as specified in the DM section
of the 'IDDO SDTM Implementation Manual'. i.e. c("DTHFL", "DTHDTC").}

\item{LB_VARS}{Specify additional variables to be included in the output
dataset. Character string. Use controlled terminology for LBTESTCD as
specified in the LB section of the 'IDDO SDTM Implementation Manual'. i.e.
c("CHOL").}

\item{MB_VARS}{Specify additional variables to be included in the output
dataset. Character string. Use controlled terminology for MBTESTCD as
specified in the MB section of the 'IDDO SDTM Implementation Manual'. i.e.
c("CRONAVIR").}

\item{MB_VARS_SPEC}{See MB_VARS above.}

\item{MP_VARS}{Specify additional variables to be included in the output
dataset. Character string. Use controlled terminology for MPLOC as
specified in the MP section of the 'IDDO SDTM Implementation Manual'. Note
this is only for LENGTH of the organ, not the WIDTH.}

\item{IN_VARS}{Specify additional variables to be included in the output
dataset. Character string. Use controlled terminology for INDECOD as
specified in the IN section of the 'IDDO SDTM Implementation Manual'. i.e.
c("PARACETAMOL").}

\item{RP_VARS}{Specify additional variables to be included in the output
dataset. Character string. Use controlled terminology for RPTESTCD as
specified in the RP section of the 'IDDO SDTM Implementation Manual'. i.e.
c("EGESTAGE", "LMPSTDTC").}

\item{SA_VARS}{Specify additional variables to be included in the output
dataset. Character string. Use controlled terminology for SADECOD as
specified in the SA section of the 'IDDO SDTM Implementation Manual'. i.e.
c("BLEEDING").}

\item{VS_VARS}{Specify additional variables to be included in the output
dataset. Character string. Use controlled terminology for VSTESTCD as
specified in the VS section of the 'IDDO SDTM Implementation Manual'. i.e.
c("MAP").}

\item{IN_DUR}{Should the analysis dataset include the duration of the
Treatment / Intervention event? This is the time from the start of the
event till the end. Boolean, default is FALSE.}

\item{IN_TIME}{Should the analysis dataset include the time since the
Treatment / Intervention event? This is the time since the end of the
event. Boolean, default is FALSE.}

\item{SA_DUR}{Should the analysis dataset include the duration of the
Clinical / Adverse Effect event? This is the time from the start of the
event till the end. Boolean, default is FALSE.}

\item{SA_TIME}{Should the analysis dataset include the time since the
Clinical / Adverse Effect event? This is the time since the end of the
event. Boolean, default is FALSE.}

\item{MP_TESTCD}{Specify which MPTESTCD is desired in the output. Options are:
"LENGTH", "WIDTH" or "BOTH". Default is Length.}
}
\value{
A dataset with one row per subject, and variables from each of the
domains listed in the parameters above.
}
\description{
Baseline analysis datasets includes events and tests conducted on the
subject's first day in the study; this also included medical history. This
function prepares, cleans, filters and pivots multiple IDDO-SDTM domains and
finally merges them into a single dataset, which can be used for research and
analysis. The choice of DISEASE preselects a number of variables which have
been chosen with input from Subject Matter Experts. The DM domain is the only
one which is required for the code to run, the rest can be optional.
}
\details{
If issues are found with the code, please log them at:
https://github.com/RhysPeploe/iddoverse/issues
}
\author{
Rhys Peploe (\code{rhys.peploe@iddo.org}, \code{rhyspeploe1998@gmail.com})
}
