% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ANALYSE_FIRST_OCCURRENCE.R
\name{ANALYSE_FIRST_OCCURRENCE}
\alias{ANALYSE_FIRST_OCCURRENCE}
\title{Create first occurrence analysis dataset.}
\usage{
ANALYSE_FIRST_OCCURRENCE(
  DISEASE_THEME = "",
  DATA_DM,
  DATA_LB = NULL,
  DATA_VS = NULL,
  DATA_IN = NULL,
  DATA_SA = NULL,
  DATA_ER = NULL,
  DATA_MB = NULL,
  DM_VARS = NULL,
  LB_VARS = NULL,
  VS_VARS = NULL,
  IN_VARS = NULL,
  SA_VARS = NULL,
  MB_VARS = NULL
)
}
\arguments{
\item{DISEASE_THEME}{The name of the disease theme being analysed. Character
string. Default is empty (selects base variables). Select from: "MALARIA",
"VL" or "EBOLA". If selection is missing or misspelt, then the base
variables will be used.}

\item{DATA_DM}{The DM domain data frame, as named in the global environment.
Required.}

\item{DATA_LB}{The LB domain data frame, as named in the global environment.}

\item{DATA_VS}{The VS domain data frame, as named in the global environment.}

\item{DATA_IN}{The IN domain data frame, as named in the global environment.}

\item{DATA_SA}{The SA domain data frame, as named in the global environment.}

\item{DATA_ER}{The ER domain data frame, as named in the global environment.}

\item{DATA_MB}{The MB domain data frame, as named in the global environment.}

\item{DM_VARS}{Specify additional variables to be included in the output
dataset. Character string. Use column names as specified in the DM section
of the 'IDDO SDTM Implementation Manual'. i.e. c("DTHFL", "DTHDTC").}

\item{LB_VARS}{Specify additional variables to be included in the output
dataset. Character string. Use controlled terminology for LBTESTCD as
specified in the LB section of the 'IDDO SDTM Implementation Manual'. i.e.
c("CHOL").}

\item{VS_VARS}{Specify additional variables to be included in the output
dataset. Character string. Use controlled terminology for VSTESTCD as
specified in the VS section of the 'IDDO SDTM Implementation Manual'. i.e.
c("MAP").}

\item{IN_VARS}{Specify additional variables to be included in the output
dataset. Character string. Use controlled terminology for INDECOD as
specified in the IN section of the 'IDDO SDTM Implementation Manual'. i.e.
c("PARACETAMOL").}

\item{SA_VARS}{Specify additional variables to be included in the output
dataset. Character string. Use controlled terminology for SADECOD as
specified in the SA section of the 'IDDO SDTM Implementation Manual'. i.e.
c("BLEEDING").}

\item{MB_VARS}{Specify additional variables to be included in the output
dataset. Character string. Use controlled terminology for MBTESTCD as
specified in the MB section of the 'IDDO SDTM Implementation Manual'. i.e.
c("CRONAVIR").}
}
\value{
A dataset with one row per subject, and variables from each of the
domains listed in the parameters above.
}
\description{
First occurrence analysis datasets include information on the first time that
an event or finding was recorded, detailing the result of the finding or
event and what day this occurred on. This function prepares, cleans, filters
and pivots multiple IDDO-SDTM domains and finally merges them into a single
dataset, which can be used for research and analysis. The choice of DISEASE
preselects a number of variables which have been chosen with input from
Subject Matter Experts. The DM domain is required.
}
\details{
If issues are found with the code, please log them at:
https://github.com/RhysPeploe/iddoverse/issues
}
\author{
Rhys Peploe
}
