% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ANALYSE_OUTCOME_VL.R
\name{ANALYSE_OUTCOME_VL}
\alias{ANALYSE_OUTCOME_VL}
\title{Create outcome analysis dataset.}
\usage{
ANALYSE_OUTCOME_VL(
  DATA_DM,
  DATA_DS = NULL,
  DATA_RS = NULL,
  DATA_MB = NULL,
  DM_VARS = NULL,
  expand_cols = FALSE
)
}
\arguments{
\item{DATA_DM}{The DM domain data frame, as named in the global environment.
Required.}

\item{DATA_DS}{The DS domain data frame, as named in the global environment.}

\item{DATA_RS}{The RS domain data frame, as named in the global environment.}

\item{DATA_MB}{The MB domain data frame, as named in the global environment.}

\item{DM_VARS}{Specify additional variables to be included in the output
dataset. Character string. Use column names as specified in the DM section
of the 'IDDO SDTM Implementation Manual'. i.e. c("AGE").}

\item{expand_cols}{Boolean option to include all DS and RS entries for each
subject. Default is FALSE, which will display the first recorded RS events
and last DS event, along with the associated VISITNUM, VISITDY and DAY. If
TRUE, all RS and DS events will be provided, along with the associated
VISITNUM, VISITDY and DAY.}
}
\value{
A dataset with one row per subject, and variables from each of the
domains listed in the parameters above.
}
\description{
Outcome analysis datasets include information about the initial and final
outcomes of studies. This function prepares, cleans, filters and pivots
multiple IDDO-SDTM domains and finally merges them into a single dataset,
which can be used for research and analysis. The choice of DISEASE preselects
a number of variables which have been chosen with input from Subject Matter
Experts. The DM, Ds and RS Domains are required, even if empty.
}
\details{
If issues are found with the code, please log them at:
https://github.com/RhysPeploe/iddoverse/issues
}
\author{
Rhys Peploe
}
