% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CLEAN_IN.R
\name{CLEAN_IN}
\alias{CLEAN_IN}
\title{Clean IN variables before preparing the domain.}
\usage{
CLEAN_IN(DATA_IN)
}
\arguments{
\item{DATA_IN}{The IN domain data frame, as named in the global environment.}
}
\value{
Data frame which has been standardised in columns: INTRT, INPRESP,
INOCCUR
}
\description{
Replaces certain character strings to standardise the INTRT column in the
Treatments and Interventions (IN) Domain.
}
\author{
Rhys Peploe
}
