% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CLEAN_MB_VL.R
\name{CLEAN_MB_VL}
\alias{CLEAN_MB_VL}
\title{Clean MB variables before preparing the domain.}
\usage{
CLEAN_MB_VL(DATA_MB)
}
\arguments{
\item{DATA_MB}{The SA domain data frame, as named in the global environment.}
}
\value{
Data frame which has been standardised in MBSTRESC and MBORRESU
}
\description{
Replaces certain character strings to standardise the MBORRESU and MBSTRESC columns in the
Microbiology (MB) Domain.
}
\author{
Rhys Peploe
}
