% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CLEAN_SA.R
\name{CLEAN_SA}
\alias{CLEAN_SA}
\title{Clean SA variables before preparing the domain.}
\usage{
CLEAN_SA(DATA_SA)
}
\arguments{
\item{DATA_SA}{The SA domain data frame, as named in the global environment.}
}
\value{
Data frame which has been standardised in columns: SATERM, SAPRESP, SAOCCUR
}
\description{
Replaces certain character strings to standardise the SATERM column in the
Clinical and Adverse Effects (SA) Domain.
}
\author{
Rhys Peploe
}
