% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_participant_map.R
\name{create_participant_map}
\alias{create_participant_map}
\title{Create a cloropleth map of the number of participants in a dataset}
\usage{
create_participant_map(DATA_DM, include_ATA = FALSE)
}
\arguments{
\item{DATA_DM}{Data frame with includes the column "COUNTRY", listing the 3
letter ISO country code where the participant is located in. The
Demographics (DM) domain in SDTM and IDDO-SDTM for example.}

\item{include_ATA}{Boolean. Should Antarctica be included on the map? Default
is \code{FALSE}}
}
\value{
A cloropleth map displaying the number of participants in each
country
}
\description{
Create a cloropleth map of the number of participants in a dataset
}
\examples{
library(worlddatr)
set.seed(123) # for reproducibility

countries <- sample(world_income$alpha_3_code, 100, replace = FALSE)

probabilities <- runif(length(countries))
probabilities <- probabilities / sum(probabilities)

country_data <- data.frame(COUNTRY = sample(countries, 10000, replace = TRUE, prob = probabilities))

create_participant_map(DATA_DM     = country_data,
                       include_ATA = FALSE)

}
\author{
Rhys Peploe
}
