% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DERIVE_AGE_DAYS.R
\name{DERIVE_AGE_DAYS}
\alias{DERIVE_AGE_DAYS}
\title{Create a variable for AGE in days.}
\usage{
DERIVE_AGE_DAYS(DATA)
}
\arguments{
\item{DATA}{Data frame containing the AGE and AGEU variables, typically the
Demographics (DM) domain.}
}
\value{
Data frame with the additional AGE_DAYS variable, providing the ages
in days.
}
\description{
Takes the AGE variable and converts the units into days. This is an
additional variable and the original AGE and AGEU will be retained when
returned. DAYS = WEEKS x 7; MONTHS x 30.417; YEARS x 365.25, result is then
rounded down.
}
\details{
AGE_DAYS is used to calculate WHO growth standards in the DERIVE_ANTHRO() function.
}
\author{
Rhys Peploe
}
