% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DERIVE_AGE_YEARS.R
\name{DERIVE_AGE_YEARS}
\alias{DERIVE_AGE_YEARS}
\title{Convert AGE to years.}
\usage{
DERIVE_AGE_YEARS(DATA)
}
\arguments{
\item{DATA}{Data frame containing the AGE and AGEU variables; typically the
Demographics (DM) domain.}
}
\value{
Data frame with AGE and AGEU in years as opposed to the original
values.
}
\description{
Convert the AGE of all subjects to years and change the AGEU to
"YEARS".
}
\author{
Rhys Peploe
}
