% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DERIVE_ANTHRO.R
\name{DERIVE_ANTHRO}
\alias{DERIVE_ANTHRO}
\title{Calculate WHO growth standards using the anthro() package.}
\usage{
DERIVE_ANTHRO(DATA)
}
\arguments{
\item{DATA}{Data frame contain the AGE, AGE_DAYS, WEIGHT & HEIGHT variables,
typically held in the Demographics (DM) and Vital Signs (VS) domains.}
}
\value{
Data frame with additional columns for WAZ, HAZ and WHZ scores,
along with flags for each.
}
\description{
Calculates the Height for Age, Weight for Age and Weight for Height z-scores
which are used to measure growth against the WHO population standards. This
is only for children less than 5 years old.
}
\details{
Note the function can be applied to a dataset with subjects greater than 5
years old too, but those subjects will have NA in these columns.
}
