% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DERIVE_SEASON.R
\name{DERIVE_SEASON}
\alias{DERIVE_SEASON}
\title{Derive the season of enrollment.}
\usage{
DERIVE_SEASON(DATA)
}
\arguments{
\item{DATA}{Data frame containing the RFSTDTC variable; typically the Demographics (DM) domain.}
}
\value{
The data frame with a SEASON column appended.
}
\description{
Creates a SEASON column based on the month in RFSTDTC with either:
WINTER (December, January, February)
SPRING (March, April, May)
SUMMER (June, July, August)
AUTUMN (September, October, November)
}
\author{
Rhys Peploe
}
