% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DERIVE_TIMING.R
\name{DERIVE_TIMING}
\alias{DERIVE_TIMING}
\title{Derive TIMING variable to subset baseline information.}
\usage{
DERIVE_TIMING(DOMAIN)
}
\arguments{
\item{DOMAIN}{Domain name in the global environment.}
}
\value{
Data frame with the new TIMING variable appended.
}
\description{
Creates a new variable 'TIMING' which takes the EPOCH variable,
and where NAs exist, they are filled with values from the VISITDY variable.
This results in TIMING being a mixture of EPOCH and VISITDY, which is used
to subset data to TIMING == "BASELINE or TIMING == 1 in order to get
baseline information from domains.
}
\author{
Rhys Peploe
}
