% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iddoverse-package.R
\docType{package}
\name{iddoverse-package}
\alias{iddoverse}
\alias{iddoverse-package}
\title{iddoverse: Transforms IDDO customised SDTM domains into analysis datasets}
\description{
Functions to convert IDDO-curated SDTM domains into analysis datasets, similar to the Analysis Data Model (ADaM).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/RhysPeploe/iddoverse}
  \item Report bugs at \url{https://github.com/RhysPeploe/iddoverse/issues}
}

}
\author{
\strong{Maintainer}: Rhys Peploe \email{rhys.peploe@iddo.org} (\href{https://orcid.org/0009-0001-1669-3716}{ORCID})

}
\keyword{internal}
