% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JOIN_HIV.R
\name{JOIN_HIV}
\alias{JOIN_HIV}
\title{Join multiple HIV variables into one combined variable for analysis.}
\usage{
JOIN_HIV(DATA, SA_TIME = FALSE, SA_DUR = FALSE)
}
\arguments{
\item{DATA}{Data frame containing both the MB test for HIV and the SA event
for HIV.}

\item{SA_TIME}{If TRUE, HIV_SAEVINTX (HIV_TIME) will be relocated after the
joint HIV variable. Boolean. Requires HIV_SAEVINTX in data frame.}

\item{SA_DUR}{If TRUE, HIV_SADUR (HIV_DUR) will be relocated after the joint
HIV variable. Boolean. Requires HIV_SADUR in data frame.}
}
\value{
Data frame with a combined HIV column and the HIV_SAOCCUR column
removed
}
\description{
HIV can be represented as a Microbiology test (MB Domain) and an Clinical and
Adverse Effect (SA Domain). This takes the MB test and where NAs exist, it is
populated by the SA entries if they exist.
}
\author{
Rhys Peploe
}
