% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JOIN_MV.R
\name{JOIN_MV}
\alias{JOIN_MV}
\title{Join multiple Malaria and VL variables into one for analysis.}
\usage{
JOIN_MV(DATA)
}
\arguments{
\item{DATA}{Data frame including --OCCUR, --DUR or --EVINTX}
}
\value{
Data frame with combined columns, with the constituent columns
removed
}
\description{
Medical History of Malaria and VL can be represented as a Clinical and
Adverse Effect (SA Domain) or as a Intervention (IN Domain). This takes
SAOCCUR and where NAs exist, they are populated with INOCCUR and renamed.
This is conducted for SADUR and INDUR, as well as SAEVINTX and INEVINTX, for
both Malaria and VL separately
}
\author{
Rhys Peploe
}
