% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PREP_APDM.R
\name{PREP_APDM}
\alias{PREP_APDM}
\title{Prepare the APDM domain for all outputs.}
\usage{
PREP_APDM(DATA_DM, DISEASE = "", VARS = NULL)
}
\arguments{
\item{DATA_DM}{The APDM domain data frame, as named in the global
environment.}

\item{DISEASE}{The name of the disease theme being analysed. Character
string. Default is NULL (selects base variables). Select from: "MALARIA",
"VL" or "EBOLA". If selection is missing or misspelt, then the base
variables will be used.}

\item{VARS}{Specify additional variables to be included in the output
dataset. Character string. Use column names as specified in the DM section
of the 'IDDO SDTM Implementation Manual'. i.e. c("DTHFL", "DTHDTC").}
}
\value{
Dataframe containing a row per USUBJID/subject.
}
\description{
Prepare the Associated Persons Demographics (APDM) Domain for
use in baseline, follow-up, treatment and outcome data sets. Takes a
IDDO-SDTM curated APDM domain and prepares the domain for merging into an
analysis data set with other domains.
}
\author{
Rhys Peploe
}
