% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PREP_DS_OUT_EVD.R
\name{PREP_DS_OUT_EVD}
\alias{PREP_DS_OUT_EVD}
\title{Prepare the DS Domain for Ebola outcome analysis.}
\usage{
PREP_DS_OUT_EVD(DATA_DS)
}
\arguments{
\item{DATA_DS}{The DS domain data frame, as named in the global environment.}
}
\value{
Dataframe containing a row per USUBJID, with DS terms as columns.
}
\description{
Prepare the Disposition (DS) domain for use in outcome analysis data sets
studying Ebola Virus Disease (EVD). Takes a IDDO-SDTM curated DS domain,
transforms and pivots it in order to present the data in a wide data format.
}
\author{
Rhys Peploe
}
