% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PREP_DS_OUT_VL.R
\name{PREP_DS_OUT_VL}
\alias{PREP_DS_OUT_VL}
\title{Prepare the DS Domain for VL outcome analysis.}
\usage{
PREP_DS_OUT_VL(DATA_DS, expand_cols = FALSE)
}
\arguments{
\item{DATA_DS}{The DS domain data frame, as named in the global environment.}

\item{expand_cols}{Boolean option to include all DS entries for each subject.
Default is FALSE, which will display the last recorded disposition event
and the associated VISITNUM, VISITDY and DAY. If TRUE, all disposition
events will be listed as DISP_NUMBER_, along with the associated VISITNUM,
VISITDY and DAY.}
}
\value{
Dataframe containing a row per USUBJID, with DS terms as columns.
}
\description{
Prepare the Disposition (DS) domain for use in outcome analysis data sets
studying Visceral Leishmaniasis. Takes a IDDO-SDTM curated DS domain,
transforms and pivots it in order to merge it into an outcome analysis data
set with other domains using the ANALYSE_OUTCOME() function.
}
\details{
All disposition records should be inspected to check the data (expand_cols =
TRUE) as events may have occurred multiple times before the final record. For
example, a subject may be lost to follow up, but have three different records
stating this, so the final disposition date may not be the actual date that
contact with the subject was lost.
}
\author{
Rhys Peploe
}
