% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PREP_ER_FIRST.R
\name{PREP_ER_FIRST}
\alias{PREP_ER_FIRST}
\title{Prepare the ER domain for analysis on the first occurrence of events.}
\usage{
PREP_ER_FIRST(DATA_ER)
}
\arguments{
\item{DATA_ER}{The ER domain data frame, as named in the global environment.}
}
\value{
Dataframe containing a row per USUBJID/subject, with ER terms and the
day of first occurrence of each as columns.
}
\description{
Extract the first occurrence of terms in the Environmental Risk (ER) domain.
Takes a IDDO-SDTM curated ER domain, transforms and pivots it in order to
merge it into a first occurrence analysis data set with other domains using
the ANALYSE_FIRST() function.
}
\author{
Rhys Peploe
}
