% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PREP_IN_B_FU.R
\name{PREP_IN_B_FU}
\alias{PREP_IN_B_FU}
\title{Further prepare the IN domain for follow up analysis, specifically for Blood
Transfusion events.}
\usage{
PREP_IN_B_FU(DATA_IN)
}
\arguments{
\item{DATA_IN}{The IN domain data frame, as named in the global environment.}
}
\value{
Dataframe containing a row per USUBJID/subject per day, with IN terms as columns.
}
\description{
Prepare the Treatments and Interventions (IN) domain for use in follow up
analysis data sets focusing on Blood Transfusions. Takes a
IDDO-SDTM curated IN domain, transforms and pivots it in order to merge it
into a follow up analysis data set with other domains using the
ANALYSE_FOLLOW_UP() function. PREP_IN_FU() and PREP_IN_B_FU() would be
merged in the ANALYSE_FOLLOW_UP() function.
}
\author{
Rhys Peploe
}
