% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PREP_IN_BL.R
\name{PREP_IN_BL}
\alias{PREP_IN_BL}
\title{Prepare the IN domain for baseline analysis.}
\usage{
PREP_IN_BL(DATA_IN, VARS = NULL, inc_DUR = FALSE, inc_TIME = FALSE)
}
\arguments{
\item{DATA_IN}{The IN domain data frame, as named in the global environment.}

\item{VARS}{Specify additional variables to be included in the output
dataset. Character string. Use controlled terminology for INDECOD as
specified in the IN section of the 'IDDO SDTM Implementation Manual'. i.e.
c("PARACETAMOL").}

\item{inc_DUR}{Should the analysis dataset include the duration of the event?
This is the time from the start of the event till the end. Boolean, default
is FALSE.}

\item{inc_TIME}{Should the analysis dataset include the time since the event?
This is the time since the end of the event. Boolean, default is FALSE.}
}
\value{
Dataframe containing a row per USUBJID, with IN terms as columns.
}
\description{
Prepare the Treatments and Interventions (IN) domain for use in baseline
analysis data sets. Takes a IDDO-SDTM curated IN domain, transforms and
pivots it in order to merge it into a baseline analysis data set with other
domains using the ANALYSE_BASELINE() function.
}
\author{
Rhys Peploe
}
