% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PREP_IN_BMV_BL.R
\name{PREP_IN_BMV_BL}
\alias{PREP_IN_BMV_BL}
\title{Further prepare the IN domain for baseline analysis, specifically for Blood
Transfusion, Malaria and VL events.}
\usage{
PREP_IN_BMV_BL(DATA_IN, inc_DUR = FALSE, inc_TIME = FALSE)
}
\arguments{
\item{DATA_IN}{The IN domain data frame, as named in the global environment.}

\item{inc_DUR}{Should the analysis dataset include the duration of the event?
This is the time from the start of the event till the end. Boolean, default
is FALSE.}

\item{inc_TIME}{Should the analysis dataset include the time since the event?
This is the time since the end of the event. Boolean, default is FALSE.}
}
\value{
Dataframe containing a row per USUBJID/subject, with IN terms as columns.
}
\description{
Prepare the Treatments and Interventions (IN) domain for use in baseline
analysis data sets focusing on Malaria and Visceral Leishmaniasis. Takes a
IDDO-SDTM curated IN domain, transforms and pivots it in order to merge it
into a baseline analysis data set with other domains using the
ANALYSE_BASELINE() function. PREP_IN_BL() and PREP_IN_BMV_BL() would be
merged in the ANALYSE_BASELINE() function.
}
\author{
Rhys Peploe
}
