% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PREP_IN_FU.R
\name{PREP_IN_FU}
\alias{PREP_IN_FU}
\title{Prepare the IN domain for follow up analysis.}
\usage{
PREP_IN_FU(DATA_IN, VARS = NULL)
}
\arguments{
\item{DATA_IN}{The IN domain data frame, as named in the global environment.}

\item{VARS}{Specify additional variables to be included in the output
dataset. Character string. Use controlled terminology for INDECOD as
specified in the IN section of the 'IDDO SDTM Implementation Manual'. i.e.
c("PARACETAMOL").}
}
\value{
Dataframe containing a row per USUBJID/subject per day, with IN terms as columns.
}
\description{
Prepare the Treatments and Interventions (IN) domain for use in follow up
analysis data sets. Takes a IDDO-SDTM curated IN domain, transforms and
pivots it in order to merge it into a follow up analysis data set with other
domains using the ANALYSE_FOLLOW_UP() function.
}
\author{
Rhys Peploe
}
