% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PREP_LB_BL.R
\name{PREP_LB_BL}
\alias{PREP_LB_BL}
\title{Prepare the LB domain for baseline analysis.}
\usage{
PREP_LB_BL(DATA_LB, DISEASE = "", VARS = NULL)
}
\arguments{
\item{DATA_LB}{The LB domain data frame, as named in the global environment.}

\item{DISEASE}{The name of the disease theme being analysed. Character
string. Default is empty (selects base variables). Select from: "MALARIA",
"VL" or "EBOLA". If selection is missing or misspelt, then the base
variables will be used.}

\item{VARS}{Specify additional variables to be included in the output
dataset. Character string. Use controlled terminology for LBTESTCD as
specified in the LB section of the 'IDDO SDTM Implementation Manual'. i.e.
c("CHOL").}
}
\value{
Dataframe containing a row per USUBJID/subject, with LBTESTCDs as
columns.
}
\description{
Prepare the Laboratory Test Results (LB) domain for use in baseline
analysis data sets. Takes a IDDO-SDTM curated LB domain, transforms and
pivots it in order to merge it into a baseline analysis data set with other
domains using the ANALYSE_BASELINE() function.
}
\author{
Rhys Peploe
}
