% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PREP_LB_FIRST.R
\name{PREP_LB_FIRST}
\alias{PREP_LB_FIRST}
\title{Prepare the LB domain for analysis on the first occurrence of events.}
\usage{
PREP_LB_FIRST(DATA_LB, DISEASE = "", VARS = NULL)
}
\arguments{
\item{DATA_LB}{The LB domain data frame, as named in the global environment.}

\item{DISEASE}{The name of the disease theme being analysed. Character
string. Default is empty (selects base variables). Select from: "MALARIA",
"VL" or "EBOLA". If selection is missing or misspelt, then the base
variables will be used.}

\item{VARS}{Specify additional variables to be included in the output
dataset. Character string. Use controlled terminology for LBTESTCD as
specified in the LB section of the 'IDDO SDTM Implementation Manual'. i.e.
c("CHOL").}
}
\value{
Dataframe containing a row per USUBJID/subject, with LBTESTCDs and the
day of first occurrence of each as columns.
}
\description{
Prepare the Laboratory (LB) domain for use in first occurrence
analysis data sets. Takes a IDDO-SDTM curated LB domain, transforms and
pivots it in order to merge it into a first occurrence analysis data set with other
domains using the ANALYSE_FIRST() function. Default variables are: "HGB", "HCT",
"PLAT", "WBC", "K", "ALT", "AST", "BILI", "CREAT", "ALB", "SODIUM", "HCG".
Disease specific features that are included by default are listed in
'Details'
}
\details{
Default Variables:

Malaria: "INTLK6", "CD4", "HGB", "HGBMET", "HCT", "PLAT", "WBC", "K", "ALT",
"AST", "BILI", "CREAT", "ALB", "SODIUM", "HCG", "G6PD"

VL: "INTLK6", "CD4", "HGB", "HCT", "PLAT", "WBC", "K", "ALT", "AST", "BILI",
"CREAT", "ALB", "SODIUM", "HCG"

Ebola: "K", "ALT", "AST", "CREAT", "SODIUM", "UREA"
}
\author{
Rhys Peploe
}
