% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PREP_LB_FU.R
\name{PREP_LB_FU}
\alias{PREP_LB_FU}
\title{Prepare the LB domain for follow up analysis.}
\usage{
PREP_LB_FU(DATA_LB, DISEASE = "", VARS = NULL)
}
\arguments{
\item{DATA_LB}{The LB domain data frame, as named in the global environment.}

\item{DISEASE}{The name of the disease theme being analysed. Character
string. Default is empty (selects base variables). Select from: "MALARIA",
"VL" or "EBOLA". If selection is missing or misspelt, then the base
variables will be used.}

\item{VARS}{Specify additional variables to be included in the output
dataset. Character string. Use controlled terminology for LBTESTCD as
specified in the LB section of the 'IDDO SDTM Implementation Manual'. i.e.
c("CHOL").}
}
\value{
Dataframe containing a row per USUBJID/subject per day, with
LBTESTCDs as columns.
}
\description{
Prepare the Laboratory Test Results (LB) domain for use in follow up analysis
data sets. Takes a IDDO-SDTM curated LB domain, transforms and pivots it in
order to merge it into a follow up analysis data set with other domains using
the ANALYSE_FOLLOW_UP() function.
}
\author{
Rhys Peploe
}
