% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PREP_MB_BL.R
\name{PREP_MB_BL}
\alias{PREP_MB_BL}
\title{Prepare the MB domain for baseline analysis.}
\usage{
PREP_MB_BL(DATA_MB, DISEASE = "", VARS = NULL)
}
\arguments{
\item{DATA_MB}{The MB domain data frame, as named in the global environment.}

\item{DISEASE}{The name of the disease theme being analysed. Character
string. Default is empty (selects base variables). Select from: "MALARIA",
"VL" or "EBOLA". If selection is missing or misspelt, then the default
variables will be used.}

\item{VARS}{Specify additional variables to be included in the output
dataset. Character string. Use controlled terminology for MBTESTCD as
specified in the MB section of the 'IDDO SDTM Implementation Manual'. i.e.
c("CRONAVIR").}
}
\value{
Data frame with one row per USUBJID/subject, with MBTESTCDs and the
units as columns
}
\description{
Prepare the Microbiology (MB) domain for use in baseline analysis data sets.
Takes a IDDO-SDTM curated MB domain, transforms and pivots it in order to
merge it into a baseline analysis data set with other domains using the
ANALYSE_BASELINE() function. Default variables are: "HIV", "AFB", "MTB",
"ANCDUOD", "ANCLMTA", "ASCLUM". Disease specific options are listed in
'Details'.
}
\details{
VL: "HIV", "AFB", "MTB", "ANCDUOD", "ANCLMTA", "ASCLUM", "PLSMDM", "PLSMDMA",
"PLSMDMS", "PFALCIP", "PFALCIPA", "PFALCIPS", "PVIVAX", "PVIVAXA", "PVIVAXS"

Ebola: "ZEBOV"
}
\author{
Rhys Peploe
}
