% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PREP_MB_FIRST.R
\name{PREP_MB_FIRST}
\alias{PREP_MB_FIRST}
\title{Prepare the MB domain for analysis on the first occurrence of events.}
\usage{
PREP_MB_FIRST(DATA_MB, DISEASE = "", VARS = NULL)
}
\arguments{
\item{DATA_MB}{The MB domain data frame, as named in the global environment.}

\item{DISEASE}{The name of the disease theme being analysed. Character
string. Default is empty (selects base variables). Select from: "MALARIA",
"VL" or "EBOLA". If selection is missing or misspelt, then the default
variables will be used.}

\item{VARS}{Specify additional variables to be included in the output
dataset. Character string. Use controlled terminology for MBTESTCD as
specified in the MB section of the 'IDDO SDTM Implementation Manual'. i.e.
c("CRONAVIR").}
}
\value{
Data frame containing a row per USUBJID/subject, with MBTESTCDs and the
day of first occurrence of each as columns.
}
\description{
Prepare the Microbiology (MB) domain for use in first occurrence
analysis data sets. Takes a IDDO-SDTM curated MB domain, transforms and
pivots it in order to merge it into a first occurrence analysis data set with other
domains using the ANALYSE_FIRST() function.
}
\author{
Rhys Peploe
}
