% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PREP_MB_FU_VL.R
\name{PREP_MB_FU_VL}
\alias{PREP_MB_FU_VL}
\title{Further prepare the MB domain for follow up analysis specifically for VL.}
\usage{
PREP_MB_FU_VL(DATA_MB)
}
\arguments{
\item{DATA_MB}{The MB domain data frame, as named in the global environment.}
}
\value{
Data frame with one row per USUBJID/subject per day, with VL specific
MBTESTCDs and the units as columns
}
\description{
Prepare the Microbiology (MB) domain for use in follow up analysis data sets
with specific actions for Visceral Leishmaniasis. Takes a IDDO-SDTM curated
MB domain, transforms and pivots it in order to merge it into a follow up
analysis data set with other domains using the ANALYSE_FOLLOW_UP() function.
PREP_MB_FU() and PREP_MB_VL_FU() would be merged in the ANALYSE_FOLLOW_UP()
function. Default variables are: "LSHMANIA", "LDONOV", "LMAJOR"
}
\author{
Rhys Peploe
}
