% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PREP_MB_BL_VL.R
\name{PREP_MB_VL_BL}
\alias{PREP_MB_VL_BL}
\title{Further prepare the MB domain for baseline analysis specifically for VL.}
\usage{
PREP_MB_VL_BL(DATA_MB)
}
\arguments{
\item{DATA_MB}{The MB domain data frame, as named in the global environment.}
}
\value{
Data frame with one row per USUBJID/subject, with VL specific
MBTESTCDs as columns
}
\description{
Prepare the Microbiology (MB) domain for use in baseline analysis data sets
with specific actions for Visceral Leishmaniasis (VL). Takes a IDDO-SDTM
curated MB domain, transforms and pivots it in order to merge it into a
baseline analysis data set with other domains using the ANALYSE_BASELINE()
function. PREP_MB_BL() and PREP_MB_VL_BL() would be merged in the
ANALYSE_BASELINE() function.
}
\author{
Rhys Peploe
}
