% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PREP_MBSPEC_FU.R
\name{PREP_MBSPEC_FU}
\alias{PREP_MBSPEC_FU}
\title{Prepare the Specimen Material Type (SPEC) and Specimen Collection Location (LOC) of follow up MB Tests.}
\usage{
PREP_MBSPEC_FU(DATA_MB, DISEASE = "", VARS = NULL)
}
\arguments{
\item{DATA_MB}{The MB domain data frame, as named in the global environment.}

\item{DISEASE}{The name of the disease theme being analysed. Character
string. Default is empty (selects base variables). Select from: "MALARIA" or
"VL". If selection is missing or misspelt, then the default
variables will be used.}

\item{VARS}{Specify additional variables to be included in the output
dataset. Character string. Use controlled terminology for MBTESTCD as
specified in the LB section of the 'IDDO SDTM Implementation Manual'. i.e.
c("CRONAVIR").}
}
\value{
Data frame with row per USUBJID/subject per day and MBTESTCDs, with MBSPEC
and MBLOC, as columns
}
\description{
Prepare the Microbiology (MB) variables MBSPEC and MBLOC for use in follow up
analysis data sets. Takes a IDDO-SDTM curated MB domain, transforms and
pivots it in order to merge it into a follow up analysis data set with other
domains using the ANALYSE_FOLLOW_UP() function.
}
\author{
Rhys Peploe
}
