% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PREP_MP_BL.R
\name{PREP_MP_BL}
\alias{PREP_MP_BL}
\title{Prepare the MP domain for baseline analysis.}
\usage{
PREP_MP_BL(DATA_MP, MPTEST = "LENGTH", VARS = NULL)
}
\arguments{
\item{DATA_MP}{The MP domain data frame, as named in the global environment.}

\item{MPTEST}{Specify which MPTESTCD is desired in the output. Options are:
"LENGTH", "WIDTH" or "BOTH". Default is Length.}

\item{VARS}{Specify additional variables to be included in the output
dataset. Character string. Use controlled terminology for MPLOC as
specified in the MP section of the 'IDDO SDTM Implementation Manual'.}
}
\value{
Dataframe containing a row per USUBJID/subject, with MPTESTCDs as
columns
}
\description{
Prepare the Morphology and Physiology (MP) domain for use in baseline
analysis data sets. Takes a IDDO-SDTM curated MP domain, transforms and
pivots it in order to merge it into a baseline analysis data set with other
domains using the ANALYSE_BASELINE() function.
}
\author{
Rhys Peploe
}
