% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PREP_MP_FU.R
\name{PREP_MP_FU}
\alias{PREP_MP_FU}
\title{Prepare the MP domain for follow up analysis.}
\usage{
PREP_MP_FU(DATA_MP, MPTEST = "LENGTH", VARS = NULL)
}
\arguments{
\item{DATA_MP}{The MP domain data frame, as named in the global environment.}

\item{MPTEST}{Specify which MPTESTCD is desired in the output. Options are:
"LENGTH", "WIDTH" or "BOTH". Default is Length.}

\item{VARS}{Specify additional variables to be included in the output
dataset. Character string. Use controlled terminology for MPLOC as
specified in the MP section of the 'IDDO SDTM Implementation Manual'.}
}
\value{
Dataframe containing a row per USUBJID/subject per day, with
MPTESTCDs and the units as columns
}
\description{
Prepare the Morphology and Physiology (MP) domain for use in follow up
analysis data sets. Takes a IDDO-SDTM curated MP domain, transforms and
pivots it in order to merge it into a follow up analysis data set with other
domains using the ANALYSE_FOLLOW_UP() function. Default variables are: "LIVER" &
"SPLEEN"
}
\author{
Rhys Peploe
}
