% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PREP_PO_FU.R
\name{PREP_PO_FU}
\alias{PREP_PO_FU}
\title{Prepare the PO domain for follow up analysis.}
\usage{
PREP_PO_FU(DATA_PO, VARS = NULL)
}
\arguments{
\item{DATA_PO}{The PO domain data frame, as named in the global environment.}

\item{VARS}{Specify additional variables to be included in the output
dataset. Character string. Use controlled terminology for POTERM as
specified in the PO section of the 'IDDO SDTM Implementation Manual'. i.e.
c("ADEVAIND").}
}
\value{
Data frame with one row per USUBJID/subject per day and POTERMs as
columns.
}
\description{
Prepare the Pregnancy Outcomes (SC) domain for use in follow up analysis data
sets. Takes a IDDO-SDTM curated PO domain, transforms and pivots it in order
to merge it into a follow up analysis data set with other domains using the
ANALYSE_FOLLOW_UP() function.
}
\author{
Rhys Peploe
}
