% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PREP_RS_OUT_VL.R
\name{PREP_RS_OUT_VL}
\alias{PREP_RS_OUT_VL}
\title{Prepare the RS Domain for VL outcome analysis.}
\usage{
PREP_RS_OUT_VL(DATA_RS, TOC_OVRLRESP = "TOC", expand_cols = FALSE)
}
\arguments{
\item{DATA_RS}{The RS domain data frame, as named in the global environment.}

\item{TOC_OVRLRESP}{Character, choice between displaying the Test of Cure
(TOC) or Overall Response (OVRLRESP). Values are thus either "TOC" or
"OVRLRESP"}

\item{expand_cols}{Boolean option to include all RS entries for each subject.
Default is FALSE, which will display the first recorded TOC or OVRLRESP
event and the associated VISITNUM, VISITDY and DAY. If TRUE, all TOCs or
OVRLRESPs will be listed as RPTESTCD_NUMBER_, along with the associated
VISITNUM, VISITDY and DAY.}
}
\value{
Dataframe containing a row per USUBJID, with RS terms as columns.
}
\description{
Prepare the Disease Response and Clinical Classification (RS) domain for use
in outcome analysis data sets studying Visceral Leishmaniasis. Takes a
IDDO-SDTM curated RS domain, transforms and pivots it in order to merge it
into an outcome analysis data set with other domains using the
ANALYSE_OUTCOME() function.
}
\author{
Rhys Peploe
}
