% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PREP_RS_OUT_VL.R
\name{PREP_RS_VL_OUT}
\alias{PREP_RS_VL_OUT}
\title{Prepare the RS Domain for VL outcome analysis.}
\usage{
PREP_RS_VL_OUT(DATA_RS)
}
\arguments{
\item{DATA_RS}{The RS domain data frame, as named in the global environment.}
}
\value{
Two functions which each create a data frame focusing on outcome
measures for VL
}
\description{
Prepare the Disease Response and Clinical Classification (RS) domain for use
in outcome analysis data sets studying Visceral Leishmaniasis. Takes a
IDDO-SDTM curated RS domain, transforms and pivots it in order to merge it
into an outcome analysis data set with other domains using the
ANALYSE_OUTCOME() function.
}
\author{
Rhys Peploe
}
