% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PREP_SA_FIRST.R
\name{PREP_SA_FIRST}
\alias{PREP_SA_FIRST}
\title{Prepare the SA domain for analysis on the first occurrence of events.}
\usage{
PREP_SA_FIRST(DATA_SA, DISEASE = "", VARS = NULL)
}
\arguments{
\item{DATA_SA}{The SA domain data frame, as named in the global environment.}

\item{DISEASE}{The name of the disease theme being analysed. Character
string. Default is empty (selects base variables). Select from: "MALARIA",
"VL" or "EBOLA". If selection is missing or misspelt, then the default
variables will be used.}

\item{VARS}{Specify additional variables to be included in the output
dataset. Character string. Use controlled terminology for SADECOD as
specified in the SA section of the 'IDDO SDTM Implementation Manual'. i.e.
c("CRONAVIR").}
}
\value{
Dataframe containing a row per USUBJID/subject, with SATERMs and the
day of first occurrence of each as columns.
}
\description{
Prepare the Clinical and Adverse Effects (SA) domain for use in first occurrence
analysis data sets. Takes a IDDO-SDTM curated SA domain, transforms and
pivots it in order to merge it into a first occurrence analysis data set with other
domains using the ANALYSE_FIRST() function.
}
\author{
Rhys Peploe
}
