% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PREP_SA_FU.R
\name{PREP_SA_FU}
\alias{PREP_SA_FU}
\title{Prepare the SA domain for follow up analysis.}
\usage{
PREP_SA_FU(DATA_SA, DISEASE = "", VARS = NULL)
}
\arguments{
\item{DATA_SA}{The SA domain data frame, as named in the global environment.}

\item{DISEASE}{The name of the disease theme being analysed. Character
string. Default is empty (selects base variables). Select from: "MALARIA",
"VL" or "EBOLA". If selection is missing or misspelt, then the default
variables will be used.}

\item{VARS}{Specify additional variables to be included in the output
dataset. Character string. Use controlled terminology for SADECOD/SATERM as
specified in the SA section of the 'IDDO SDTM Implementation Manual'. i.e.
c("CRONAVIR").}
}
\value{
Data frame with one row per USUBJID/subject per day, with SATERMs as
columns
}
\description{
Prepare the Clinical and Adverse Effects (SA) domain for use in follow up
analysis data sets. Takes a IDDO-SDTM curated SA domain, transforms and
pivots it in order to merge it into a follow up analysis data set with other
domains using the ANALYSE_FOLLOW_UP() function.
}
\author{
Rhys Peploe
}
