% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PREP_SA_MV_BL.R
\name{PREP_SA_MV_BL}
\alias{PREP_SA_MV_BL}
\title{Further prepare the SA domain for baseline analysis specifically for Malaria
and VL variables.}
\usage{
PREP_SA_MV_BL(DATA_SA, inc_DUR = FALSE, inc_TIME = FALSE)
}
\arguments{
\item{DATA_SA}{The SA domain data frame, as named in the global environment.}

\item{inc_DUR}{Should the analysis dataset include the duration of the event?
This is the time from the start of the event till the end. Boolean, default
is FALSE.}

\item{inc_TIME}{Should the analysis dataset include the time since the event?
This is the time since the end of the event. Boolean, default is FALSE.}
}
\value{
Data frame with one row per USUBJID/subject, with Malaria and VL
specific SATERMs as columns
}
\description{
Prepare the Clinical and Adverse Effects (SA) domain for use in baseline
analysis data sets focussing on Malaria and Visceral Leishmaniasis (VL).
Takes a IDDO-SDTM curated SA domain, transforms and pivots it in order to
merge it into a baseline analysis data set with other domains using the
ANALYSE_BASELINE() function. PREP_SA_BL() and PREP_SA_MV_BL() would be merged
in the ANALYSE_BASELINE() function.
}
\author{
Rhys Peploe
}
