% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PREP_SC_FU.R
\name{PREP_SC_FU}
\alias{PREP_SC_FU}
\title{Prepare the SC domain for follow up analysis.}
\usage{
PREP_SC_FU(DATA_SC, VARS = NULL)
}
\arguments{
\item{DATA_SC}{The SC domain data frame, as named in the global environment.}

\item{VARS}{Specify additional variables to be included in the output
dataset. Character string. Use controlled terminology for SCTESTCD as
specified in the SC section of the 'IDDO SDTM Implementation Manual'. i.e.
c("ADEVAIND").}
}
\value{
Wide data frame with one row per USUBJID/subject per day, with
SCTESTCDs as columns.
}
\description{
Prepare the Subject Characteristics (SC) domain for use in follow up analysis
data sets. Takes a IDDO-SDTM curated SC domain, transforms and pivots it in
order to merge it into a follow up analysis data set with other domains using
the ANALYSE_FOLLOW_UP() function.
}
\author{
Rhys Peploe
}
