% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PREP_TS.R
\name{PREP_TS}
\alias{PREP_TS}
\title{Prepare the TS domain for all outputs}
\usage{
PREP_TS(DATA_TS, DATA_DM)
}
\arguments{
\item{DATA_TS}{The TS domain data frame, as named in the global environment.}

\item{DATA_DM}{The DM domain data frame, as named in the global environment.}
}
\value{
Data frame with a row per STUDYID, and ARMCD (treatment arm code),
DISEASE and STUDYFU (study follow up) as columns.
}
\description{
Prepare the Trial Summary (TS) domain for use in any analysis data sets.
Takes a IDDO-SDTM curated TS domain, subsets and joins it with the
Demographics (DM) domain, which can then be merged into an analysis data set
with other domains.
}
\author{
Rhys Peploe
}
