% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PREP_VS_FU.R
\name{PREP_VS_FU}
\alias{PREP_VS_FU}
\title{Prepare the VS domain for follow up analysis.}
\usage{
PREP_VS_FU(DATA_VS, DISEASE = "", VARS = NULL)
}
\arguments{
\item{DATA_VS}{The VS domain data frame, as named in the global environment.}

\item{DISEASE}{The name of the disease theme being analysed. Character
string. Default is empty (selects base variables). Select from: "MALARIA",
"VL" or "EBOLA". If selection is missing or misspelt, then the base
variables will be used.}

\item{VARS}{Specify additional variables to be included in the output
dataset. Character string. Use controlled terminology for VSTESTCD as
specified in the VS section of the 'IDDO SDTM Implementation Manual'. i.e.
c("MAP").}
}
\value{
Data frame with one row per USUBJID/subject, with VSTESTCDs as
columns
}
\description{
Prepare the Vital Signs (VS) domain for use in follow up
analysis data sets. Takes a IDDO-SDTM curated VS domain, transforms and
pivots it in order to merge it into a follow up analysis data set with other
domains using the ANALYSE_FOLLOW_UP() function.
}
\author{
Rhys Peploe
}
