% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PREP_VS_SCR_TEMP_BL.R
\name{PREP_VS_SCR_TEMP_BL}
\alias{PREP_VS_SCR_TEMP_BL}
\title{Prepare the Temperature variable in the VS domain for baseline analysis,
using 'Screening' as the timing filter.}
\usage{
PREP_VS_SCR_TEMP_BL(DATA_VS)
}
\arguments{
\item{DATA_VS}{The VS domain data frame, as named in the global environment.}
}
\value{
Data frame with one row per USUBJID/subject, with TEMP, TEMP_UNITS
and TEMP_LOC as columns
}
\description{
Prepare the Temperature variable in the Vital Signs (VS) domain for use in
baseline analysis data sets. Instead of the typical TIMING == 1 or BASELINE,
this takes EPOCH = SCREENING as the definition of baseline. Takes a IDDO-SDTM
curated VS domain, filters just the temperature records, transforms and
pivots it in order to merge it into a baseline analysis data set with other
domains using the ANALYSE_BASELINE() function.
}
\details{
This allows the Temperature location to be recorded in the analysis dataset
without including the location for all of the other VS variables
}
\author{
Rhys Peploe
}
