% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PREP_VS_TEMP_FU.R
\name{PREP_VS_TEMP_FU}
\alias{PREP_VS_TEMP_FU}
\title{Prepare the Temperature variable in the VS domain for follow up analysis.}
\usage{
PREP_VS_TEMP_FU(DATA_VS)
}
\arguments{
\item{DATA_VS}{The VS domain data frame, as named in the global environment.}
}
\value{
Data frame with one row per USUBJID/subject, with TEMP, TEMP_UNITS
and TEMP_LOC as columns
}
\description{
Prepare the Temperature variable in the Vital Signs (VS) domain for use in
follow up analysis data sets. Takes a IDDO-SDTM curated VS domain, filters
just the temperature records, transforms and pivots it in order to merge it
into a follow up analysis data set with other domains using the
ANALYSE_FOLLOW_UP() function.
}
\details{
This allows the Temperature location to be recorded in the analysis dataset
without including the location for all of the other VS variables
}
\author{
Rhys Peploe
}
