% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RP_RPTESTB.R
\docType{data}
\name{RP_RPTESTB}
\alias{RP_RPTESTB}
\title{IDDO-SDTM Tutorial Reproductive System Findings (RP) Domain}
\format{
\subsection{}{

A data frame with 1 row and 18 columns:
\describe{
\item{STUDYID}{Study Identifier}
\item{DOMAIN}{Domain Abbreviation}
\item{RPSEQ}{Sequence Number}
\item{USUBJID}{Unique Subject Identifier}
\item{RPTESTCD}{Short Name of Reproductive Test}
\item{RPTEST}{Name of Reproductive Test}
\item{RPORRES}{Result or Finding in Original Units}
\item{RPORRESU}{Original Units}
\item{RPMODIFY}{Pre-Specified Event}
\item{RPSTRESC}{Character Result/Finding in Std Format}
\item{RPSTRESN}{Numeric Result/Finding in Standard Units}
\item{RPSTRESU}{Standard Units}
\item{VISITNUM}{Visit Number}
\item{VISIT}{Visit Name}
\item{VISITDY}{Planned Study Day of Visit}
\item{EPOCH}{Epoch}
\item{RPDTC}{Date/Time of Collection}
\item{RPDY}{Study Day of Visit/Collection/Exam}
}
}
}
\usage{
RP_RPTESTB
}
\description{
A fictional RP dataset for testing code and training
}
\keyword{datasets}
