% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SA_RPTESTB.R
\docType{data}
\name{SA_RPTESTB}
\alias{SA_RPTESTB}
\title{IDDO-SDTM Tutorial Clinical and Adverse Events (SA) Domain}
\format{
\subsection{}{

A data frame with 3 rows and 21 columns:
\describe{
\item{STUDYID}{Study Identifier}
\item{DOMAIN}{Domain Abbreviation}
\item{SASEQ}{Sequence Number}
\item{USUBJID}{Unique Subject Identifier}
\item{SATERM}{Reported Term for the Event}
\item{SAMODIFY}{Modified Reported Term}
\item{SADECOD}{Dictionary-Derived Term}
\item{SACAT}{Category for Event}
\item{SAPRESP}{Pre-Specified Event}
\item{SAOCCUR}{Clinical Event Occurrence}
\item{SASEV}{Severity/Intensity}
\item{VISITNUM}{Visit Number}
\item{VISIT}{Visit Name}
\item{VISITDY}{Planned Study Day of Visit}
\item{EPOCH}{Epoch}
\item{SADY}{Study Day of Specimen Collection}
\item{SADTC}{Date/Time of Specimen Collection}
\item{SASTDY}{Study Day of Start of Event}
\item{SAENDY}{Study Day of End of Event}
\item{SADUR}{Duration}
\item{SAEVINTX}{Evaluation Interval Text}
}
}
}
\usage{
SA_RPTESTB
}
\description{
A fictional SA dataset for testing code and training
}
\keyword{datasets}
