% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TS_RPTESTB.R
\docType{data}
\name{TS_RPTESTB}
\alias{TS_RPTESTB}
\title{IDDO-SDTM Tutorial Trial Summary (TS) Domain}
\format{
\subsection{}{

A data frame with 25 rows and 11 columns:
\describe{
\item{STUDYID}{Study Identifier}
\item{DOMAIN}{Domain Abbreviation}
\item{TSSEQ}{Sequence Number}
\item{TSGRPID}{Group ID}
\item{TSPARMCD}{Trial Summary Parameter Short Name}
\item{TSPARM}{Trial Summary Parameter}
\item{TSVAL}{Parameter Value}
\item{TSVALNF}{Parameter Null Flavor}
\item{TSVALCD}{Parameter Value Code}
\item{TSVCDREF}{Name of the Reference Terminology}
\item{TSVCDVER}{Version of the Reference Terminology}
}
}
}
\usage{
TS_RPTESTB
}
\description{
A fictional TS dataset for testing code and training
}
\keyword{datasets}
