% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VS_RPTESTB.R
\docType{data}
\name{VS_RPTESTB}
\alias{VS_RPTESTB}
\title{IDDO-SDTM Tutorial Vital Signs (VS) Domain}
\format{
\subsection{}{

A data frame with 18 row and 25 columns:
\describe{
\item{STUDYID}{Study Identifier}
\item{DOMAIN}{Domain Abbreviation}
\item{USUBJID}{Unique Subject Identifier}
\item{VSSEQ}{Sequence Number}
\item{VSREFID}{Reference ID}
\item{VSTESTCD}{Vital Signs Test Short Name}
\item{VSTEST}{Vital Signs Test Name}
\item{VSCAT}{Category for Vital Signs}
\item{VSSCAT}{}
\item{VSORRES}{Result or Finding in Original Units}
\item{VSORRESU}{Original Units}
\item{VSMODIFY}{Pre-Specified Event}
\item{VSSTRESC}{Character Result/Finding in Std Format}
\item{VSSTRESN}{Numeric Result/Finding in Standard Units}
\item{VSSTRESU}{Standard Units}
\item{VSSTAT}{Completion Status}
\item{VSREASND}{Reason Not Performed}
\item{VSLOC}{Location of Vital Signs Measurement}
\item{VISITNUM}{Visit Number}
\item{VISIT}{Visit Name}
\item{VISITDY}{Planned Study Day of Visit}
\item{EPOCH}{Epoch}
\item{VSDTC}{Date/Time of Measurements}
\item{VSDY}{Study Day of Vital Signs}
\item{VSTPT}{Planned Time Point Name}
}
}
}
\usage{
VS_RPTESTB
}
\description{
A fictional VS dataset for testing code and training
}
\keyword{datasets}
