% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/worlddatr-package.R
\docType{package}
\name{worlddatr-package}
\alias{worlddatr}
\alias{worlddatr-package}
\title{worlddatr: A Package For Global Choropleth Maps}
\description{
Data and functions for working with country and territory data including ISO codes, World Bank income groups, longitude and latitude.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://infectious-diseases-data-observatory.github.io/worlddatr/}
}

}
\author{
\strong{Maintainer}: Rhys Peploe \email{rhyspeploe1998@gmail.com} (\href{https://orcid.org/0009-0001-1669-3716}{ORCID})

Authors:
\itemize{
  \item Rhys Peploe \email{rhyspeploe1998@gmail.com} (\href{https://orcid.org/0009-0001-1669-3716}{ORCID})
}

}
\keyword{internal}
