% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_country_to_iso.R
\name{convert_country_to_iso}
\alias{convert_country_to_iso}
\title{Convert country names to ISO country codes}
\usage{
convert_country_to_iso(data, country_name_col)
}
\arguments{
\item{data}{Data frame with a column of country names}

\item{country_name_col}{Character. The name of the column containing the
country names.}
}
\value{
The input data frame with two new columns alpha_3_code and
alpha_2_code representing the ISO standardised 3 & 2 letter country codes
}
\description{
Matches a column of country names in your data with a bank of various country
names (\code{country_name_lookup.xlsx} in \code{inst/extdata}) and appends two new
columns for the ISO 3 and 2 letter codes in order to use standardised codes
instead of messy country names. The function reports in the R console how
many rows did not get matched with the data bank.
}
\examples{
country_names = data.frame(country = sample(world_income$country, 100))

convert_country_to_iso(country_names, "country")

}
