% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_participant_map.R
\name{create_participant_map}
\alias{create_participant_map}
\title{Create a cloropleth map of the number of participants in a dataset}
\usage{
create_participant_map(
  DATA_DM,
  include_ATA = FALSE,
  include_n = TRUE,
  title = "",
  subtitle = "",
  legend = "Number of Participants",
  colour_high = "#14B1E7",
  colour_low = "#CCECF9",
  colour_default = "#FDF3F4",
  colour_borders = "black",
  colour_background = "#FFFFFF",
  scale_breaks = pretty_breaks(),
  log_scale = FALSE
)
}
\arguments{
\item{DATA_DM}{Data frame with includes the column "COUNTRY", listing the 3
letter ISO country code where the participant is located in. The
Demographics (DM) domain in SDTM and IDDO-SDTM for example.}

\item{include_ATA}{Boolean. Should Antarctica be included on the map? Default
is \code{FALSE}.}

\item{include_n}{Boolean. Should \code{n = X} where X is the number of
participants, be included in the map. Default is \code{TRUE}.}

\item{title}{Character. Title of the choropleth map.}

\item{subtitle}{Character. Subtitle of the choropleth map.}

\item{legend}{Character. Legend title of the choropleth map.}

\item{colour_high}{Colour of the high end of the scale, either using the name
of a base colour or the HEX code, i.e. "red" or "#F9250C. This will be the
colour of the country with the highest value. Default is IDDO-branded blue
"#14B1E7".}

\item{colour_low}{Colour of the low end of the scale, either using the name
of a base colour or the HEX code, i.e. "red" or "#F9250C. This will be the
colour of the country with the lowest value. Default is IDDO branded blue
at 30\% tint, "#CCECF9".}

\item{colour_default}{Colour of the other countries without participant data,
either using the name of a base colour or the HEX code, i.e. "red" or
"#F9250C. Default is IDDO branded red at 5\% tint, "#FDF3F4".}

\item{colour_borders}{Colour of the country borders, either using the name of
a base colour or the HEX code, i.e. "red" or "#F9250C. Default is "black".}

\item{colour_background}{Colour of the plot background, either using the name
of a base colour or the HEX code, i.e. "red" or "#F9250C". Default is
"#FFFFFF" (white).}

\item{scale_breaks}{Numeric list. Specify custom scale breaks for the legend,
i.e. c(100, 500, 1000, 2000). Default is pretty_breaks from the scales
package.}

\item{log_scale}{Boolean. Should the number of participants be transformed
into the log scale for when the distribution of values is very uneven.
Default is \code{FALSE}.}
}
\value{
A cloropleth map displaying the number of participants in each
country
}
\description{
Create a cloropleth map of the number of participants in a dataset
}
\examples{
library(worlddatr)
set.seed(123) # for reproducibility

countries <- sample(world_income$alpha_3_code, 100, replace = FALSE)

probabilities <- runif(length(countries))
probabilities <- probabilities / sum(probabilities)

country_data <- data.frame(COUNTRY = sample(countries, 10000, replace = TRUE, prob = probabilities))

create_participant_map(DATA_DM     = country_data,
                       include_ATA = FALSE,
                       include_n   = TRUE,
                       title       = "Number of Participants in Studies across Countries",
                       subtitle    = "Data from studies registered between 2000 - 2024",
                       colour_high = "#14B1E7",
                       colour_low  = "#CCECF9",
                       colour_default = "#FDF3F4",
                       colour_borders = "black",
                       colour_background = "#FFFFFF",
                       scale_breaks = pretty_breaks(),
                       log_scale = FALSE
                       )

}
\author{
Rhys Peploe
}
