% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JOIN_PREGNANT.R
\name{JOIN_PREGNANT}
\alias{JOIN_PREGNANT}
\title{Join multiple pregnancy indicator variables into one combined variable for
analysis.}
\usage{
JOIN_PREGNANT(DATA)
}
\arguments{
\item{DATA}{Data frame containing both the MB test for HIV and the SA event
for HIV.}
}
\value{
Data frame with a PREGNANT column, and PREGIND & HCG removed.
}
\description{
Pregnancy can be confirmed with a lab test (LB Domain) or by asking the
subject (RP Domain). This takes the LB test column and where NAs exist, they
are populated with the result of PREGIND, the pregnancy indicator from the RP
domain.
}
\author{
Rhys Peploe
}
