% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/worlddatr-package.R
\docType{package}
\name{worlddatr-package}
\alias{worlddatr}
\alias{worlddatr-package}
\title{worlddatr: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
