% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/worlddatr-package.R
\docType{package}
\name{worlddatr-package}
\alias{worlddatr}
\alias{worlddatr-package}
\title{worlddatr: A package for global spatial data analysis and choropleth maps}
\description{
Data and functions for working with country and territory data including ISO codes, World Bank income groups, longitude and latitude.
}
\author{
\strong{Maintainer}: Rhys Peploe \email{rhys.peploe@iddo.org} (\href{https://orcid.org/0009-0001-1669-3716}{ORCID})

}
\keyword{internal}
